move and rename fetch_file_from_url
